global date state
