+++ components/Calendar/Calendar.tsx
@@ -1,0 +1,12 @@
+import { NextPage } from 'next'
+// import FullCalendar from '@fullcalendar/react'
+// import dayGridPlugin from '@fullcalendar/daygrid'
+
+const Calendar: NextPage = () => {
+  // This causes internal server error.
+  // return <FullCalendar plugins={[dayGridPlugin]} initialView="dayGridMonth" />
+
+  return <div>Calendar</div>
+}
+
+export default Calendar
