+++ package.json
@@ -9,6 +9,8 @@
     "test": "jest --watch"
   },
   "dependencies": {
+    "@fullcalendar/daygrid": "^5.10.1",
+    "@fullcalendar/react": "^5.10.1",
     "@types/cookies": "^0.7.7",
     "@types/styled-components": "^5.1.15",
     "axios": "^0.24.0",
